# backend/api-docs.yaml
openapi: 3.0.0
info:
  title: Miner Node Hardware Comparison API
  version: 1.0.0
  description: API for comparing hardware requirements across different blockchain miner nodes

servers:
  - url: http://localhost:8080
    description: Development server

paths:
  /nodes:
    get:
      summary: List all nodes
      description: Retrieve a list of all miner nodes
      responses:
        "200":
          description: List of nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /nodes/{id}:
    get:
      summary: Get node by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Node ID
      responses:
        "200":
          description: Node details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  schemas:
    Node:
      type: object
      properties:
        id:
          type: integer
        blockchain_type:
          type: string
        cpu_cores:
          type: integer
        ram_gb:
          type: integer
        storage_gb:
          type: integer
        network_mbps:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - blockchain_type
        - cpu_cores
        - ram_gb
        - storage_gb
        - network_mbps

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
                example: "Not Found"
              message:
                type: string

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              type:
                type: string
                example: "Internal Server Error"
              message:
                type: string